refactor: ContentPage component
