+++ lib/types.ts
@@ -1,3 +1,5 @@
+import { PortableTextBlockComponent } from "@portabletext/react"
+
 export type Geopoint = {
   lat: number
   lng: number
@@ -97,4 +99,10 @@ export type CarouselItem = Pick<Entry, "dates" | "location" | "name" | "slug">
 export type EntryType = {
   title: string
   value: string
+}
+
+export type Page = {
+  slug?: string
+  title?: string
+  content?: any[]
 }
