+++ pages/licence.tsx
@@ -1,23 +1,12 @@
 import { trpc } from "@/lib/trpc"
 import NoopLayout from "../components/layout/NoopLayout"
-import { PortableText } from '@portabletext/react'
-import { pageComponents } from "../lib/page"
+import { ContentPage } from "@/components/ContentPage"
 
 const LicencePage = () => {
-  const { data: page } = trpc.page.useQuery({ pageSlug: "licence"})
+  const { data: licensePage } = trpc.page.useQuery({ pageSlug: "licence"})
 
   return (
-    <div className="w-1/3">
-      <h1 className="text-5xl font-semibold mb-12">
-        { page?.title }
-      </h1>
-      <div className="flex flex-col">
-        <PortableText
-          value={page?.content}
-          components={pageComponents}
-        />
-      </div>
-    </div>
+    <ContentPage page={licensePage} />
   )
 }
